+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Added option to remove all "unused" component items through recursion
 - Improved the openapi-format terminal log output and how verbose is handled
+- Remove markdown comments (syntax [comment]: <>) from the description fields 
 
 ## [1.5.0] - 2020-10-05